Python Game Bot - Input Automation
YoloV5-based game script that relies upon several custom-built scripts to detect objects on-screen, interact with the game UI, and perform cyclic activities in a logically controlled manner. 

Built in an OOP format for easy implementation and updating
Highlights flexibility and functionality of object detection for application manipulation and automation.
Highlights UI management capabilities using computer vision techniques.

There are only a few issues and potential improvements I have currently detected with this implementation:

- Average model accuracy. There are certain cosmetics that the model reads as wisps. This is done by the player base to prevent botting efforts, and is only partially effective. Requires an improved model to overcome these images.
- Advanced UI management. Can be updated to include the ability to right click and rotate the screen to find objects. This has been implemented as standalone functionality in other scripts, but can be tied together with more advanced, global, modularized UI Management function calls, which will ultimately require a rewrite of the entire backup_ui_management script. This will take time. 
- Some static variable definitions
- Some UX issues
- No remote monitoring
- Not entirely accurate inventory counts
- Minimal exception handling (detecting 'training' and 'emptying' poses to verify the action is occurring).

There are perhaps a few more. But for now, this stands as the current implementation.